update(product): check permission

+++ src/product/ProductController.java
@@ -1,5 +1,7 @@
 package product;
+
 import etc.Response;
+import exceptions.ForbiddenException;
 import config.Database;
 import utils.HttpUtil;
 import product.dtos.SearchProductDTO;
@@ -8,7 +10,7 @@ import product.dtos.DeleteProductDTO;
 import product.dtos.UpdateProductDTO;
 
 import java.io.IOException;
-import java.io.PrintWriter; 
+import java.io.PrintWriter;
 import java.io.BufferedReader;
 import java.lang.RuntimeException;
 import java.sql.SQLException;
@@ -21,19 +23,20 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 @WebServlet("/product/*")
 public class ProductController extends HttpServlet {
+	private ProductService productService = new ProductService();
+	private HttpUtil util = new HttpUtil();
+
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException {
-		ProductService productService = new ProductService();
 		String route = (req.getPathInfo() != null) ? req.getPathInfo().substring(1) : "";
 		Gson gson = new Gson();
 		try {
 			switch (route) {
 				case "list":
-					// TODO: use SearchProductDTO
-					HttpUtil<SearchProductDTO> util = new HttpUtil();
 					SearchProductDTO dto = util.getBodyContentFromReq(req, SearchProductDTO.class);
 					List<ProductEntity> products = productService.getAll(getServletContext(), dto);
 
@@ -69,12 +72,22 @@ public class ProductController extends HttpServlet {
 		try {
 			switch (route) {
 				case "":
-					HttpUtil<CreateProductDTO> util = new HttpUtil();
-					CreateProductDTO dto= util.getBodyContentFromReq(req, CreateProductDTO.class);
+					HttpSession session = req.getSession();
+					Integer role = (Integer) session.getAttribute("role");
+					if (role == null) {
+						throw new ForbiddenException("Please login and try again!");
+					}
+					if (role != 2) {
+						throw new ForbiddenException("Not enough permission!");
+					}
+
+					HttpUtil util = new HttpUtil();
+					CreateProductDTO dto = util.getBodyContentFromReq(req, CreateProductDTO.class);
 					productService.createOne(getServletContext(), dto);
 
 					res.setContentType("application/json;charset=UTF-8");
-					Response<Object> response = new Response(200, "Create a product successfully!!", null);
+					Response<Object> response = new Response(200, "Create a product successfully!!",
+							null);
 					try (PrintWriter out = res.getWriter()) {
 						out.println(gson.toJson(response));
 					}
@@ -92,8 +105,15 @@ public class ProductController extends HttpServlet {
 			try (PrintWriter out = res.getWriter()) {
 				out.println(gson.toJson(response));
 			}
+		} catch (ForbiddenException e) {
+			Response<Object> response = new Response(403, e.getMessage(), null);
+			res.setContentType("application/json");
+			try (PrintWriter out = res.getWriter()) {
+				out.println(gson.toJson(response));
+			}
 		}
 	}
+
 	@Override
 	protected void doDelete(HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException {
 		ProductService productService = new ProductService();
@@ -102,12 +122,22 @@ public class ProductController extends HttpServlet {
 		try {
 			switch (route) {
 				case "":
-					HttpUtil<DeleteProductDTO> util = new HttpUtil();
+					HttpSession session = req.getSession();
+					Integer role = (Integer) session.getAttribute("role");
+					if (role == null) {
+						throw new ForbiddenException("Please login and try again!");
+					}
+					if (role != 2) {
+						throw new ForbiddenException("Not enough permission!");
+					}
+
+					HttpUtil util = new HttpUtil();
 					DeleteProductDTO dto = util.getBodyContentFromReq(req, DeleteProductDTO.class);
 					productService.deleteOne(getServletContext(), dto);
 
 					res.setContentType("application/json;charset=UTF-8");
-					Response<Object> response = new Response(200, "Delete a product successfully!!", null);
+					Response<Object> response = new Response(200, "Delete a product successfully!!",
+							null);
 					try (PrintWriter out = res.getWriter()) {
 						out.println(gson.toJson(response));
 					}
@@ -121,6 +151,12 @@ public class ProductController extends HttpServlet {
 			}
 		} catch (RuntimeException e) {
 			Response<Object> response = new Response(500, e.getMessage(), null);
+			res.setContentType("application/json");
+			try (PrintWriter out = res.getWriter()) {
+				out.println(gson.toJson(response));
+			}
+		} catch (ForbiddenException e) {
+			Response<Object> response = new Response(403, e.getMessage(), null);
 			res.setContentType("application/json");
 			try (PrintWriter out = res.getWriter()) {
 				out.println(gson.toJson(response));
@@ -136,12 +172,22 @@ public class ProductController extends HttpServlet {
 		try {
 			switch (route) {
 				case "":
-					HttpUtil<UpdateProductDTO> util = new HttpUtil();
+					HttpSession session = req.getSession();
+					Integer role = (Integer) session.getAttribute("role");
+					if (role == null) {
+						throw new ForbiddenException("Please login and try again!");
+					}
+					if (role != 2) {
+						throw new ForbiddenException("Not enough permission!");
+					}
+
+					HttpUtil util = new HttpUtil();
 					UpdateProductDTO dto = util.getBodyContentFromReq(req, UpdateProductDTO.class);
 					productService.updateOne(getServletContext(), dto);
 
 					res.setContentType("application/json;charset=UTF-8");
-					Response<Object> response = new Response(200, "Delete a product successfully!!", null);
+					Response<Object> response = new Response(200, "Delete a product successfully!!",
+							null);
 					try (PrintWriter out = res.getWriter()) {
 						out.println(gson.toJson(response));
 					}
@@ -155,6 +201,12 @@ public class ProductController extends HttpServlet {
 			}
 		} catch (RuntimeException e) {
 			Response<Object> response = new Response(500, e.getMessage(), null);
+			res.setContentType("application/json");
+			try (PrintWriter out = res.getWriter()) {
+				out.println(gson.toJson(response));
+			}
+		} catch (ForbiddenException e) {
+			Response<Object> response = new Response(403, e.getMessage(), null);
 			res.setContentType("application/json");
 			try (PrintWriter out = res.getWriter()) {
 				out.println(gson.toJson(response));
